rename domain methods
