+++ components/common/RouteGuard.tsx
@@ -10,7 +10,7 @@ import { useSpaces } from 'hooks/useSpaces';
 import { useUser } from 'hooks/useUser';
 import { filterSpaceByDomain } from 'lib/spaces/filterSpaceByDomain';
 import { redirectToAppLogin, shouldRedirectToAppLogin } from 'lib/utilities/browser';
-import { getValidCustomDomain } from 'lib/utilities/domains/getValidCustomDomain';
+import { getCustomDomainFromHost } from 'lib/utilities/domains/getCustomDomainFromHost';
 // Pages shared to the public that don't require user login
 // When adding a page here or any new top-level pages, please also add this page to DOMAIN_BLACKLIST in lib/spaces/config.ts
 const publicPages = ['/', 'share', 'api-docs', 'u', 'join', 'invite', 'authenticate', 'test'];
@@ -74,7 +74,7 @@ export default function RouteGuard({ children }: { children: ReactNode }) {
 
     // condition: no user session and no wallet address
     else if (!user) {
-      if (getValidCustomDomain()) {
+      if (getCustomDomainFromHost()) {
         // if app is running on a custom domain, main url will handle login
         return { authorized: true };
       }
