+++ lib/utilities/domains/getSpaceDomainFromUrlPath.ts
