+++ lib/utilities/getAppOriginURL.ts
@@ -2,7 +2,7 @@ import { appSubdomain, isDevEnv, isStagingEnv } from 'config/constants';
 import { getAppApexDomain } from 'lib/utilities/domains/getAppApexDomain';
 import { isLocalhostAlias } from 'lib/utilities/domains/isLocalhostAlias';
 
-import { getValidSubdomain } from './getValidSubdomain';
+import { getValidSubdomain } from './domains/getSpaceDomainFromHost';
 
 export function getAppOriginURL({ protocol, port, host }: { protocol?: string; host?: string; port?: string }) {
   const appDomain = getAppApexDomain();
